Manifest database framework and implementation
OpTool will utilize a database to track usages of OpTool and opping/deoping players. A database handler will be implemented for connecting to an external database.

Config file is now modified to include new connection strings to connect to the database.

Updated build to add the dev branch for compile tests.